requirements.txt
tensorflow==2.15.0
keras==2.15.0
opencv-python==4.8.1.78
flask==3.0.0
pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.4
werkzeug==3.0.1
gunicorn==21.2.0 